Move oauth.go into auth_login.go, rename helpers to *_helper.go
