+++ internal/cmd/completion_helper.go
